fix(bazel): fix macos support for bazel build (#377)
This fixes the use of Linux-only flags on a `cp` command in the Bazel rules. It removes the use of `-s`, which just makes symbolic links of everything instead of copying, removes the use of `--preserve=links` which preserves symbolic links in the source, and changes `-r` to `-R` which should do the same thing as `--preserve=links` in retaining symbolic links.