chore: update googleapis and goldens (#388)
Update the `googleapis` hash used in the repo to match the submodule hash that is checked in. As a result the golden files for integration tests that depend on the version of `googleapis` defined in the Bazel environment must be updated to reflect any proto changes as a result of the `googleapis` update.

This update is also necessary to get a sufficiently recent version of `googleapis` that contains the `google/cloud/extended_operations.proto`.