fix: enable ignore_unknown in config file parsers (#390)
This just ignores the unknown fields in a JSON/YAML blob when parsing config files. Specifically, the `ServiceYamlConfig.php` doesn't have a `type` field (which it doesn't need), but this field may appear in the YAML files. At some point, the default was changed in `google/protobuf` and in order to upgrade (see #385) we need to explicitly set `ignore_unknown = true`.