fix: catalog proto service to proto file (#399)
Add a mapping of `ServiceDescriptorProto` to `FileDescriptorProto` to `ProtoCatalog`. This is especially useful when a reference to a `ServiceDescriptorProto` is available, and a `FileDescriptorProto` is necessary in order to convert that into `ServiceDetails`. This is essentially a parent file lookup.

This is necessary for DIREGAPIC LRO support, because the "Resource" service client needs to be able to configure the "Operations" service client named in the annotations.