fix: remove extraneous message in custom lro tests (#410)
Generated unit tests for the exception case of Custom LRO methods contained an extraneous request message object and its construction. The variable was never used after it was constructed because the test focuses on asserting that an exception happens. The existing LRO exception tests don't use it either. Might want to circle back in a separate PR to clean up that one too.