chore(ast): support array, vector, handle map (#506)
Support "codifying" `array` and `Vector` types by wrapping them in the corresponding `AST` expression, just like how primitives are handled. Specifically handle `Map` rather than having it be reported as `object`. Include the type of the element that can't be handled in the exception.

With this change, we no longer need to wrap every `array` or `Vector` in an `AST::array` or convert `vector->toArray()` then wrap it - this is done by the AST. Makes code generation helpers easier to test by not requiring that AST Expressions be passed around between function calls, saving code generation/AST wrapping until necessary.